+++ plugins/permission/constants.ts
@@ -2,7 +2,16 @@ export const ECDSA_SIGNER_CONTRACT =
     "0x6A6F069E2a08c2468e7724Ab3250CdBFBA14D4FF"
 export const WEBAUTHN_SIGNER_CONTRACT =
     "0x8AA55d4BfAE101609078681A69B5bc3181516612"
-export const CALL_POLICY_CONTRACT = "0xe4Fec84B7B002273ecC86baa65a831ddB92d30a8"
+export const CALL_POLICY_CONTRACT_V5_3_1 =
+    "0xe4Fec84B7B002273ecC86baa65a831ddB92d30a8"
+/**
+ * @dev CALL_POLICY_CONTRACT_V5_3_2 updates
+ * - Added ONE_OF condition
+ * - Added native ETH transfer support
+ * - Added `zeroAddress` target address support, which means you can approve any contracts with specific selector. (e.g. approve any ERC20 transfer)
+ */
+export const CALL_POLICY_CONTRACT_V5_3_2 =
+    "0x09dc5ca0d9e5bd4330b45b40585126519d82b6cc" // TODO: need to confirm the address after review
 export const GAS_POLICY_CONTRACT = "0xaeFC5AbC67FfD258abD0A3E54f65E70326F84b23"
 export const RATE_LIMIT_POLICY_CONTRACT =
     "0xf63d4139B25c836334edD76641356c6b74C86873"