+++ plugins/permission/policies/callPolicyUtils.ts
@@ -8,14 +8,21 @@ import {
     isHex,
     pad,
     toFunctionSelector,
-    toHex
+    toHex,
+    Address,
+    AbiParameter
 } from "viem"
-import type {
-    CombinedArgs,
-    GeneratePermissionFromArgsParameters,
-    ParamRule,
-    PermissionCore
+import {
+    GeneratePermissionWithPolicyAddressParameters,
+    ParamCondition,
+    type CombinedArgs,
+    type ParamRule,
+    type PermissionCore
 } from "./types.js"
+import {
+    CALL_POLICY_CONTRACT_V5_3_1,
+    CALL_POLICY_CONTRACT_V5_3_2
+} from "../constants.js"
 
 export function getPermissionFromABI<
     TAbi extends Abi | readonly unknown[],
@@ -23,8 +30,9 @@ export function getPermissionFromABI<
 >({
     abi,
     args,
-    functionName
-}: GeneratePermissionFromArgsParameters<TAbi, TFunctionName>): Pick<
+    functionName,
+    policyAddress
+}: GeneratePermissionWithPolicyAddressParameters<TAbi, TFunctionName>): Pick<
     PermissionCore,
     "selector" | "rules"
 > {
@@ -46,10 +54,16 @@ export function getPermissionFromABI<
     let paramRules: ParamRule[] = []
     if (args && Array.isArray(args)) {
         paramRules = (args as CombinedArgs<AbiFunction["inputs"]>)
-            .map(
-                (arg, i) =>
-                    arg && {
-                        param: pad(
+            .map((arg, i) => {
+                if (!arg) return null
+                if (policyAddress === CALL_POLICY_CONTRACT_V5_3_1) {
+                    if (arg.condition === ParamCondition.ONE_OF) {
+                        throw Error(
+                            "ONE_OF condition is not supported in CALL_POLICY_CONTRACT_V5_3_1"
+                        )
+                    }
+                    return {
+                        params: pad(
                             isHex(arg.value)
                                 ? arg.value
                                 : toHex(
@@ -60,7 +74,41 @@ export function getPermissionFromABI<
                         offset: i * 32,
                         condition: arg.condition
                     }
-            )
+                }
+                if (policyAddress === CALL_POLICY_CONTRACT_V5_3_2) {
+                    let params: Hex[]
+                    if (arg.condition === ParamCondition.ONE_OF) {
+                        params = arg.value.map((value) =>
+                            pad(
+                                isHex(value)
+                                    ? value
+                                    : toHex(
+                                          value as Parameters<typeof toHex>[0]
+                                      ),
+                                { size: 32 }
+                            )
+                        )
+                    } else {
+                        params = [
+                            pad(
+                                isHex(arg.value)
+                                    ? arg.value
+                                    : toHex(
+                                          arg.value as Parameters<
+                                              typeof toHex
+                                          >[0]
+                                      ),
+                                { size: 32 }
+                            )
+                        ]
+                    }
+                    return {
+                        params,
+                        offset: i * 32,
+                        condition: arg.condition
+                    }
+                }
+            })
             .filter((rule) => rule) as ParamRule[]
     }
     return {
@@ -70,7 +118,8 @@ export function getPermissionFromABI<
 }
 
 export const encodePermissionData = (
-    permission: PermissionCore | PermissionCore[]
+    permission: PermissionCore | PermissionCore[],
+    policyAddress: Address
 ): Hex => {
     const permissionParam = {
         components: [
@@ -102,10 +151,12 @@ export const encodePermissionData = (
                         name: "offset",
                         type: "uint64"
                     },
-
                     {
-                        name: "param",
-                        type: "bytes32"
+                        name: "params",
+                        type:
+                            policyAddress === CALL_POLICY_CONTRACT_V5_3_2
+                                ? "bytes32[]"
+                                : "bytes32"
                     }
                 ],
                 name: "rules",